Move uploads to weapons manager
